useradd-staticids.bbclass: make missing uid a fatal error
When USERADD_ERROR_DYNAMIC is enabled, packages are silently
skipped if they attempt to add a user/group which do not have
a static uid/gid.  If the bitbake recipe depends on one of the
skipped packages, the build will fail without indicating that
the reason it failed was due to a missing uid/gid.  This change
makes skipping packages a fatal error and prints a message to
that effect.

Signed-off-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Rich Tollerton <rich.tollerton@ni.com>
Acked-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>
Natinst-ReviewBoard-ID: 75780