util-linux: break out util-linux-getopt
The getopt binary itself is only a few kb, while the util-linux package
is much larger. The lsb package is very small and depends only on
getopt. Break it out into util-linux-getopt in order to reduce the disk
requirements of lsb on systems that don't otherwise bring in util-linux.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com
Acked-by: Brad Mouring <brad.mouring@ni.com>
Natinst-CAR-ID: 497617
Natinst-Reviewboard-ID: 81494

Conflicts:
	meta/recipes-core/util-linux/util-linux.inc