tar: Fix bug with --selinux option and unlabeled files
When SELinux is enabled in the kernel but no policy is loaded, files may
be marked as unlabeled.  When these files are processed,
rpl_lgetfilecon() returns the security context as "unlabeled".
map_to_failure() then frees the security context, sets errno to ENODATA,
and returns -1.  However, since the security context is not NULL,
xattr_selinux_coder() attempts to read from it when the header is
generated, which leads to memory corruption (and a failure on some
future malloc).

For unlabeled files, set the security context to NULL to avoid this
use-after-free bug.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Josh Cartwright <joshc@ni.com>
Reviewed-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 91520
Natinst-CAR-ID: 517518