udev-cache: Fix handling of path names with special characters
By default GNU tar unquotes the file names read. That creates problems with
devices under /dev that contain special characters or things that look like
escape sequences when unquoted e.g:
  "/dev/ni/NiRioSrv:fpga:00010000\nirio_transport\0"

Additionally to compound the problem the busybox version of tar does not
unquote by default and does not support the GNU options for specifying the
quoting style.

Change the tar invocation to pass the '--no-unquote' option via TAR_OPTIONS
environment variable (doing it this way allows us to support both the busybox
tar and the GNU tar).

Also as an optimization eliminate the call to 'cut' to strip the leading '/'
by invoking tar in '/' with 'dev' as a relative path (which stops tar from
complaining about the leading '/', the reason the 'cut' call was there in the
fist place).

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
---
Natinst-CAR-ID: 523592
Natinst-ReviewBoard-ID: 95960