base-files: check tty before setting mesg
The default profile runs 'mesg n' to disable write access to the
terminal.  However, if this is run when stdin is not a tty, it generates
a spurious message:

stdin: is not a tty

To fix this, check that stdin is a tty before running 'mesg n'.  Since
busybox's version of mesg does not support the '-s' (silent) flag,
redirect the output of 'tty' to /dev/null.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Ken Sharp <ken.sharp@ni.com>
Reviewed-by: Brad Mouring <brad.mouring@ni.com>
Reviewed-by: Scot Salmon <scot.salmon@ni.com>
Reviewed-by: Rich Tollerton <rich.tollerton@ni.com>
---
Natinst-CAR-ID: 450463
Natinst-ReviewBoard-ID: 98649