PACKAGE_INSTALL_NODEPS: add
PACKAGE_INSTALL (via IMAGE_INSTALL) is the normative way of adding
packages to images. However, it is not generally compatible with
BUILD_IMAGES_FROM_FEEDS, because RDEPENDS+="${PACKAGE_INSTALL}" in
image.bbclass; this makes it impossible to add packages to an image that
only exist in IPK form, not recipe form.

To ameliorate this situation, let's create a new variable
PACKAGE_INSTALL_NODEPS, which is basically the same as PACKAGE_INSTALL,
except it explicitly skips all dependency checks.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-Reviewboard-ID: 110700
Reviewed-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>
Reviewed-by: Haris Okanovic <haris.okanovic@ni.com>