pt-sans: use ni.com mirror of pt-sans .zip file
The upstream source of pt-sans at http://www.fontstock.com/public/PTSansOFL.zip
does not have any version information or revision history. At some point
in 2014 it appears they changed the contents of the file, including the
sizes of the .ttf files and name of the license file. We could update
the recipe with new checksums pretty easily, but there's no guarantee we
wouldn't need to do it again, and it also means that images built after
that update have different pt-sans .ttf files than those built before,
which is a reproducibility issue. Also, the fontstock.com site does not
inspire confidence, with no https support and several very dated or
entirely broken pages, and no documentation of why the .zip file was
changed. Continuing to use the one we already shipped, even if we have
to host it ourselves, seems like a more robust solution all around.

Signed-off-by: Scot Salmon <scot.salmon@ni.com>