isoimage-isohybrid.py: add grub smbios module
Grub by default does not understand smbios tables unless the 'smbios'
module is installed and loaded. wic uses grub to generate a native efi
boot image which is monolithic, thus we have to build in the smbios
module inside it to support reading from smbios while provisioning from
a usb stick for example.

This commit is NOT upstreamable until smbios support is added to grub,
either via patches in the OE upstream grub recipe or upstream grub
integrates this feature and OE updates the recipe past the integration
point.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>
(this commit is not in origin/master)