linux-firmware: package carl9170 separately
carl9170 is an atheros-based firmware, the succesor of ar9170 which is
deprecated and its driver (ar9170) was removed from the kernel tree.
carl9170 comes both as a binary blob and also with free sources; the
sources are deleted in linux-firmware.bb to avoid depeding on bash/etc.

Package the carl9170 binary firmware separately because it is needed
by various usb dongles and installing whole linux-firmware is overkill.

Also we should keep ar9170 as is because OOT drivers might still use it.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>
(this commit is not in origin/master)