libsolv: enable MULTI_SEMANTICS
By default, libsolv uses the rpm logic for version comparison, which is
not quite the same as debian. Opkg now sets the distribution type for
libsolv to be debian. But for that to work, libsolv needs to be compiled
with MULTI_SEMANTICS=ON.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(cherry picked from commit 8ba4fb906bbcd01c36871c3fccc50975821c0de4)