base-files: install /etc/skel files into ROOT_HOME
Home directory default files stored in /etc/skel ought to be installed
for the root user as well. However, because /etc/skel is primarily
utilized by useradd, which is never invoked for root (for the good
reason that root must already exist), root winds up never getting a
default .profile or .bashrc.

We fix this by explicitly installing .profile and .bashrc into
ROOT_HOME. This is analogous to how Fedora solved the problem [1],
except instead of installing them in the package "filesystem" (which is
their base-files), they use a separate package "rootfiles" for this sort
of stuff, to separate the ancillary root support files from core
OS/filesystem structure. Since we haven't yet made that distinction,
insofar as base-files owns /etc/skel/.bashrc, instead of bash, we are in
no deeper trouble than we already are by letting base-files install
/root/.bashrc too.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=999114

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-Reviewboard-ID: 171298