kmod: gimp depmodwrapper-cross
* Comment out the exit call if kernel abiversion mismatches the
requested kernel version. While this logic is reasonable for other OE
builds, it will cause breakages in the NILRT build workflow when the
package feed is build in a different environment from the image
recipes.

! This patch papers over a deeper problem with our build workflow, but
one which is already being solved. This commit can be dropped when we
move to a unified build environment!

Signed-off-by: Alex Stewart <alex.stewart@ni.com>