autogen-native: fix shebang too long error
OE upstream removed autogen-native starting with the Rocko release, so
this commit will be dropped after rebasing on Rocko (don't upstream it).

Because we combine NIBuild with OE in older NILRT thus combining their
paths we end up in a situation where shebangs are too long (over the
standard 128 bytes default in Linux kernels). Fix autogen-native by
relying on /usr/bin/env instead of hardcoding the path.

Upstream-Status: Not-Upstreamable [NI-specific, to be dropped post Rocko]

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>