site/*-linux: don't cache ac_cv_sizeof_bool
The value was hardcoded from the time it couldn't be computed, which
is no longer the case. After C99 'bool' is only defined if stdbool.h
is included, it's implementation defined and not required to be 1, so
caching it doesn't make sense and makes certain recipes whose code
tests ac_cv_sizeof_bool fail to build.

Natinst-CAR-ID: 679977

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>
(This commit is not in origin/master, I've sent it but it's a P0 CAR
so I can't wait after upstream - Ionel)