lsbinitscripts: replace dependency with initd-functions
Commits 1202307b24 ("lsbinitscripts: don't use update-alternatives")
cdcebd81c87 ("initscripts: don't use update-alternatives") and
061fa614cec ("update-alternatives.bbclass: refuse to manage SysV init
scripts") make sure that the lsbinitscripts and initscripts conflict
with eachother and can't be alternatives on target devices.

lsb has a hardcoded dependency on lsbinitscripts which makes it
impossible to install on systems using initscripts and lsbinitscripts
in turn has a runtime dependency on the full util-linux package,
making it an unwanted dependency on busybox systems.

Therefore make lsb depend on initd-functions, which lsbinitscripts
RPROVIDES, allowing for multiple versions of /etc/init.d/functions and
installing lsb depending on initscripts without the full util-linux.

The default initd-functions provider is initscripts, so this changes
the dependency of lsb (but no the runtime behaviour AFAIK). I did not
change the default provider because sysvinit depends on initscripts
being the default.

The old dependency, previous to this change, can be brought back with
PREFERRED_RPROVIDER_initd-functions = "lsbinitscripts".

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>