HACK: glibc-locale: fix 2.24 locale build breakage
IMO fixing the host build of binary locales is hopeless with all these
2.24 reverts because the cross-localedef tool is broken, xlocale.h was
dropped from glibc starting with 2.26 and now >=2.26 is expected by all
distros released in 2018 onwards.

Just build the locales on target on first boot, it's instant, a few
miliseconds max.

This commit should be dropped once we upgrade past glibc 2.24.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>