xinput_calibrator: Add support for libinput
Current implementation only supports Evdev, Usbtouchscreen or standard
Xorg calibrators. When libinput is installed, the standard Xorg
calibrator ends up being used but the generated conf file does not
work. This patch adds support for libinput by adding a new calibrator
so that xinput_calibrator works even when libinput is installed.

This is based on code from Evdev.cpp and Evdev.hpp with the following
changes
  o Removed evdev specific calibration code and output formats
  o Added dynamic calibration code for libinput
  o Added suport for xinput and xorg.conf.d output types

Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@ni.com>
Natinst-ReviewBoard-ID: 329820
Acked-by: Alex Stewart <alex.stewart@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>